refactor(cmcd): extract shared reporter internals into modules
Move copyReportValues/copyItemValue, createCmcdReporterConfig,
createEncodingOptions, the normalized config type aliases, and
CMCD_REQUIRED_EVENT_KEYS out of CmcdReporter.ts into their own
internal modules. No behavior change; public API surface (index.ts,
cml-cmcd.api.md) is untouched.
